Refactoring in RandomizedShrinkablesGenerator
